Init project last version

+++ Output/src/Timer.cpp
@@ -1,13 +1,10 @@
 // ----------------------------------------------------
-// Timer with milisecons precision
+// Fast timer with milisecons precision
 // ----------------------------------------------------
 
 #include "Timer.h"
 #include "SDL2\SDL_timer.h"
-
-// L2: TODO 1: Fill Start(), Read(), ReadSec() methods
-// they are simple, one line each!
-
+	
 Timer::Timer()
 {
 	Start();
@@ -25,5 +22,5 @@ int Timer::ReadSec() const
 
 float Timer::ReadMSec() const
 {
-	return SDL_GetTicks() - startTime;
+	return (float)(SDL_GetTicks() - startTime);
 }
